Add test configuration to Docker Compose
This commit sets up the configuration for running tests in the Docker Compose YAML file. Specifically, it defines a service, “tests”, which will run using the Dockerfile.dev image. It’s also set up to use the same volume definitions as the web service for consistency, and runs the "npm run test" command on startup.