Add Dockerfile to build and serve the app
A new Dockerfile is added in this commit to build and serve our Node.js application using a 2-stage Docker build. In the first stage, the application is built using node:16-alpine and in the second stage, the built files are served using nginx. This change was needed to deliver a Docker container that can be easily deployed and run in any environment with Docker.