Add Deployment Workflow and Update Docker Configuration
Introduced a deploy.yaml workflow to automate build and deployment process on GitHub Actions. The workflow is triggered every time a new commit is pushed to the main branch. It builds the various Docker images and deploys them to Dockerhub, before deploying to AWS Elastic Beanstalk. Also, modified the docker-compose.yaml to use the production Dockerfile and open on port 80 instead of 3000. Finally, a new docker-compose.dev.yaml file was created to handle development environment separately, enabling it to run tests and keeping the app on port 3000. This ensures a clear separation of concerns between production and development environments, and aids in ensuring the production build is as lean and optimal as possible.